Move blaze::step_function to separate file
